requirements.txt
bytewax==0.15.1
requests==2.28.0
kafka-python==2.0.2
